Add files via upload
Our project is to predict the cultivation of crop based on high volume of production by predicting the weather.The fields in the dataset are StateName ,DistrictName,CropYear,Season,Area,Production and Crops.The field Production is taken as Class variables .The Dataset achieves the accuracy of about 81.30%. The idea is based on giving Crops and Production as input. The algorithm DecisionTree Classifier is used to decide on whether the crop can be cultivated or not in a particular season . If the Production value of the given crop exceeds the mean value of the Production for the corresponding season, then the crop can be cultivated. If the condition fails, it gives an alert  that the crop cannot be cultivated. Thus it helps the farmers to decide whether they can cultivate the crops on a particular season.